# perf: optimize LowercaseBytes with loop unrolling