@thevilledev 2025-03-24
Hey @xiaost, thanks for your feedback. Indeed, it's used for URI scheme and host canonicalisation. Nevertheless occurs on every request that's being processed and I guess the HTTP header-like data is a relevant benchmark regardless? Would another benchmark be needed?