@thevilledev 2025-03-24
I did a short string focused benchmark based on the one mentioned on the title. There's actually some performance regression. I think it outweighs the benefits. I'll close off the PR and come back to this later. Thanks for the review help anyway!