@xiaost 2025-03-24
@thevilledev Never mind. Thanks again.

considering the func is short enough that can be inlined it's good for current cases I think.